requirements.txt
python-dotenv
requests
tabulate